safari-534.53-branch/LayoutTests/plugins/netscape-destroy-plugin-script-objects-expected.txt

CONSOLE MESSAGE: line 0: PLUGIN: NPP_Destroy
This tests that accessing plug-in script objects from a destroyed plugin cause throws the right exceptions.
SUCCESS
